# fix: schema compatibility fingerprint