--- a/codec.go
+++ b/codec.go
@@ -34,8 +34,7 @@ type ValEncoder interface {
 
 // ReadVal parses Avro value and stores the result in the value pointed to by obj.
 func (r *Reader) ReadVal(schema Schema, obj any) {
-	key := cacheFingerprintOf(schema)
-	decoder := r.cfg.getDecoderFromCache(key, reflect2.RTypeOf(obj))
+	decoder := r.cfg.getDecoderFromCache(schema.CacheFingerprint(), reflect2.RTypeOf(obj))
 	if decoder == nil {
 		typ := reflect2.TypeOf(obj)
 		if typ.Kind() != reflect.Ptr {
@@ -66,15 +65,14 @@ func (w *Writer) WriteVal(schema Schema, val any) {
 
 func (c *frozenConfig) DecoderOf(schema Schema, typ reflect2.Type) ValDecoder {
 	rtype := typ.RType()
-	key := cacheFingerprintOf(schema)
-	decoder := c.getDecoderFromCache(key, rtype)
+	decoder := c.getDecoderFromCache(schema.CacheFingerprint(), rtype)
 	if decoder != nil {
 		return decoder
 	}
 
 	ptrType := typ.(*reflect2.UnsafePtrType)
 	decoder = decoderOfType(c, schema, ptrType.Elem())
-	c.addDecoderToCache(key, rtype, decoder)
+	c.addDecoderToCache(schema.CacheFingerprint(), rtype, decoder)
 	return decoder
 }