When a schema contains a compatibility change anywhere in its tree, the entire schema should show the change in its fingerprint. To make this simpler, the writer fingerprint is carried in the schema, to show this change. There is an edge case in records that the exact defaults must be accounted for in the cached fingerprint.

This has the benefit of slightly optimising the planning phase.
```
name                      old time/op    new time/op    delta
SuperheroDecode-8            232ns ± 1%     229ns ± 1%  -1.27%  (p=0.000 n=10+10)
SuperheroEncode-8            198ns ± 0%     199ns ± 1%  +0.68%  (p=0.000 n=10+10)
PartialSuperheroDecode-8     106ns ± 0%      97ns ± 1%  -8.05%  (p=0.000 n=8+9)
SuperheroGenericDecode-8    2.17µs ± 0%    2.17µs ± 0%    ~     (p=0.376 n=8+9)
SuperheroGenericEncode-8     251ns ± 2%     247ns ± 0%  -1.67%  (p=0.000 n=10+8)
SuperheroWriteFlush-8        162ns ± 1%     162ns ± 2%    ~     (p=0.675 n=9+10)

name                      old alloc/op   new alloc/op   delta
SuperheroDecode-8            47.0B ± 0%     47.0B ± 0%    ~     (all equal)
SuperheroEncode-8             112B ± 0%      112B ± 0%    ~     (all equal)
PartialSuperheroDecode-8     9.00B ± 0%     9.00B ± 0%    ~     (all equal)
SuperheroGenericDecode-8    2.04kB ± 0%    2.04kB ± 0%    ~     (all equal)
SuperheroGenericEncode-8     80.0B ± 0%     80.0B ± 0%    ~     (all equal)
SuperheroWriteFlush-8        0.00B          0.00B         ~     (all equal)

name                      old allocs/op  new allocs/op  delta
SuperheroDecode-8             0.00           0.00         ~     (all equal)
SuperheroEncode-8             1.00 ± 0%      1.00 ± 0%    ~     (all equal)
PartialSuperheroDecode-8      0.00           0.00         ~     (all equal)
SuperheroGenericDecode-8      60.0 ± 0%      60.0 ± 0%    ~     (all equal)
SuperheroGenericEncode-8      3.00 ± 0%      3.00 ± 0%    ~     (all equal)
SuperheroWriteFlush-8         0.00           0.00         ~     (all equal)
```