@redaLaanait 2024-02-07
That's a considerable improvement! The last time, it seems I missed the big picture regarding the cache key's logic. 

Following a similar reasoning, we should also ensure that the resolved composite schema can't be a child node in a tree (that has a normal schema as a root node)?